:recycle: rename component names
