+++ src/components/card-slider/card-slider-card.vue
