+++ src/components/submit-loader-button/submit-loader-button.vue
